allow for garbage collection algo preference
https://theindiestone.com/forums/index.php?/topic/47417-dedicated-server-on-linux-vps-jre-crashes/

Only way to keep my server stable with 10+ people is to set to G1GC, but there's other java jre gc options

Afaik there's only a few java GC's that'd matter. ZGC is usually best, but G1GC can sometimes be more efficient with it's RAM usage. Probably related to something dumb in either my vm configs or otherwise, but seems like an easy change to allow it to be configurable more generally.

Maybe just adds tech debt to potential future work to deal with #38 and #23 but idk, this is an easy fix for my specific problems.

Signed-off-by: Dylan Shepard <dylan@shepard.dev>